Moving LLM test to another file to isolate its environment
